pyproject.toml
[project]
name = "telegram-mcp"
version = "0.1.0"
description = "MCP server for Telegram chat search"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "telethon>=1.34",
    "mcp>=0.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "tomli>=2.0",
    "aiofiles>=23.0",
    "filelock>=3.0",
    "pypdf>=4.0",
    "pytesseract>=0.3",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "pre-commit>=3.6",
]

[project.scripts]
telegram-mcp = "src.mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram", "src/storage", "src/search", "src/mcp", "src/extraction"]

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
